erase unuse hooks
