+++ package.json
@@ -36,35 +36,23 @@
     "@testing-library/user-event": "^13.2.1",
     "axios": "^0.27.2",
     "classnames": "^2.3.1",
-    "dayjs": "^1.11.2",
-    "dotenv": "^16.0.1",
-    "http-proxy-middleware": "^2.0.6",
-    "i18n-iso-countries": "^7.4.0",
-    "i18next": "^21.8.0",
-    "i18next-browser-languagedetector": "^6.1.4",
     "lodash": "^4.17.21",
     "node-sass": "^7.0.1",
     "react": "^18.1.0",
     "react-dom": "^18.1.0",
     "react-error-boundary": "^3.1.4",
-    "react-i18next": "^11.16.9",
-    "react-paypal-express-checkout": "^1.0.5",
     "react-query": "^3.39.0",
     "react-redux": "^8.0.1",
-    "react-router-dom": "6",
+    "react-responsive": "^9.0.0-beta.8",
     "react-scripts": "5.0.1",
     "react-use": "^17.3.2",
-    "recoil": "^0.7.3-alpha.2",
-    "store": "^2.0.12",
     "web-vitals": "^2.1.0"
   },
   "devDependencies": {
-    "@types/jest": "^27.0.1",
     "@types/lodash": "^4.14.182",
     "@types/node": "^16.7.13",
     "@types/react": "^18.0.0",
     "@types/react-dom": "^18.0.0",
-    "@types/store": "^2.0.2",
     "babel-eslint": "^10.1.0",
     "eslint": "^8.14.0",
     "eslint-config-airbnb": "^19.0.4",
