+++ src/components/MobileSearchList/MobileSearchList.module.scss
