+++ src/components/SearchList/index.tsx
@@ -42,7 +42,6 @@ const SearchList = ({ isLoading }: Props) => {
     setIndex(Number(e.currentTarget.dataset.idx));
   };
 
-  // TODO 처음 키 두 번 입력됨
   useEffect(() => {
     document.addEventListener('keydown', handleKeyPress);
 
@@ -66,7 +65,6 @@ const SearchList = ({ isLoading }: Props) => {
             onMouseEnter={handleMouseEnter}
           >
             <SearchIcon className={styles.icon} />
-            {/* <span>{item.sickNm}</span> */}
             <span>
               {item.sickNm.split(',').map((letter, i) => {
                 const key = `${item.sickCd}-${i}`;
