--- src/hooks/worker/index.tsx
@@ -1,2 +1,0 @@
-export { useAxios, CancelToken, isCancel, instance as axios, plainAxios } from './useAxios';
-export type { AxiosError, AxiosResponse } from './useAxios';
