+++ src/pages/MobileSearchDiseases/MobileSearchDiseases.module.scss
