+++ src/pages/SearchDiseases/SearchDiseases.module.scss
