--- src/utils/locale/countries.ts
@@ -1,8 +1,0 @@
-import i18nCountries from 'i18n-iso-countries'
-import en from 'i18n-iso-countries/langs/en.json'
-import ko from 'i18n-iso-countries/langs/ko.json'
-
-i18nCountries.registerLocale(en)
-i18nCountries.registerLocale(ko)
-
-export default i18nCountries
