--- src/utils/locale/en/front.json
@@ -1,5 +1,0 @@
-{
-	"city.kwangmyung": "Kwangmyung",
-	"gnb.todo": "Todo",
-	"gnb.weather": "Weather"
-}
