--- src/utils/locale/ko/front.json
@@ -1,5 +1,0 @@
-{
-	"city.kwangmyung": "광명시",
-	"gnb.todo": "할일",
-	"gnb.weather": "날씨"
-}
